@@ -5,13 +5,13 @@
     "author": "Mikhail Nepomniashchii",
     "dependencies": {
         "@chakra-ui/react": "^1.8.8",
-        "@emotion/react": "^11.7.1",
-        "@emotion/styled": "^11.6.0",
-        "framer-motion": "^6.2.3",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
-        "react-icons": "^3.0.0",
-        "react-scripts": "^5.0.0"
+        "@emotion/react": "^11.9.0",
+        "@emotion/styled": "^11.8.1",
+        "framer-motion": "^6.3.1",
+        "react": "^18.0.0",
+        "react-dom": "^18.0.0",
+        "react-icons": "^4.3.1",
+        "react-scripts": "^5.0.1"
     },
     "scripts": {
         "start": "react-scripts start",
@@ -36,10 +36,10 @@
         ]
     },
     "devDependencies": {
-        "@types/node": "^17.0.18",
-        "@types/react": "^17.0.39",
-        "@types/react-dom": "^17.0.11",
-        "typescript": "^4.5.5",
-        "vercel": "^23.1.2"
+        "@types/node": "^17.0.25",
+        "@types/react": "^18.0.6",
+        "@types/react-dom": "^18.0.2",
+        "typescript": "^4.6.3",
+        "vercel": "^24.1.0"
     }
 }
