@@ -1,20 +1,16 @@
-import { ColorModeScript } from "@chakra-ui/react"
-import * as React from "react"
-import ReactDOM from "react-dom"
-import { App } from "./components/App"
-import * as serviceWorker from "./serviceWorker"
+import * as React from 'react';
+import { createRoot } from 'react-dom/client';
+import { ColorModeScript } from '@chakra-ui/react';
 
-import './index.css'
+import { App } from './components/App';
 
-ReactDOM.render(
-  <React.StrictMode>
+import './index.css';
+
+const root = createRoot(document.getElementById('root') as Element);
+
+root.render(
+  <>
     <ColorModeScript />
     <App />
-  </React.StrictMode>,
-  document.getElementById("root"),
-)
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://cra.link/PWA
-serviceWorker.unregister()
+  </>,
+);
